Refactors build process for watch mode
Updates the build process to use esbuild's context API for
better watch mode support. This improves the handling of file
changes and ensures that CSS and module packaging are correctly
updated during development. Also updates the DCC sheet to use the
new TABS and PARTS API rather than renderInner.

Adds a script for symlinking the module for local development
convenience.